ENH: revisit simulation method, setting up with make_glbl
